Forbid deleting Resourse and MeasurementUnit if it is using in ReceiptResource table

+++ src/Api/AutomatedWarehouse.Api.Infrastructure/Services/Guide services/MeasurementUnitService.cs
@@ -1,4 +1,5 @@
-﻿using AutomatedWarehouse.Api.Domain.Models;
+﻿using AutomatedWarehouse.Api.Domain.Exceptions;
+using AutomatedWarehouse.Api.Domain.Models;
 using AutomatedWarehouse.Api.Infrastructure.Database;
 using Microsoft.EntityFrameworkCore;
 
@@ -24,8 +25,11 @@ namespace AutomatedWarehouse.Api.Infrastructure.Services.Guide_services
             await context.SaveChangesAsync();
         }
 
-        public async Task DeleteAsync(Guid entityId) //check
+        public async Task DeleteAsync(Guid entityId)
         {
+            if (context.ReceiptResources.Any(x => x.MeasurementUnitId == entityId))
+                throw new EntityIsInUsageException("Cannot delete MeasurementUnit with current Id because it is currently using in another table");
+
             var measurementUnit = await context.MeasurementUnits.SingleAsync(x => x.Id == entityId);
             context.MeasurementUnits.Remove(measurementUnit);
             await context.SaveChangesAsync();
